pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "django-account"
description = "Accountverwaltung in Django"
readme = "README.md"
authors = [{ name = "Rafael Urben", email = "github@rafaelurben.ch" }]
maintainers = [{ name = "Rafael Urben", email = "github@rafaelurben.ch" }]
license = { file = "LICENSE" }
classifiers = [
    "Private :: Do not upload - Too specific for reuse without modification",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">= 3.10"
dependencies = [
    "Django>=4.2, <6",
    "social-auth-app-django>=5.0.0, <6",
    "django-passkeys>=1.2.7, <2",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
]

[project.urls]
Homepage = "https://rafaelurben.github.io/django-account"
Docs = "https://rafaelurben.github.io/django-account"
GitHub = "https://github.com/rafaelurben/django-account"
Issues = "https://github.com/rafaelurben/django-account/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages]
find = { namespaces = false }

[tool.setuptools.dynamic]
version = { file = ["_version.txt"] }

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
